==> notebooks/2.0-dictionary-based-approach.ipynb <==
import pandas as pd
import re

# Define the threshold for toxicity
TOXICITY_THRESHOLD = 0.8
DATA_PATH = '../data/'
# Load the dataset
df = pd.read_csv(DATA_PATH + 'raw/filtered.tsv', delimiter='\t')
df.head()

# Build a dictionary of toxic phrases and their replacements
toxic_dict = {}
for index, row in df.iterrows():
    if row['ref_tox'] > TOXICITY_THRESHOLD:
        toxic_dict[row['reference'].lower()] = row['translation'].lower()

# Helper function to maintain the same case
def replace_with_same_case(match_obj):
    match_str = match_obj.group(0)
    replacement = toxic_dict[match_str.lower()]
    if match_str[0].isupper():
        return replacement.capitalize()
    return replacement

# Function to detoxify text
def detoxify(text):
    pattern = re.compile('|'.join(re.escape(key) for key in toxic_dict.keys()), re.IGNORECASE)
    return pattern.sub(replace_with_same_case, text)

# Test the function with a new sentence
test_sentence = "I like that shit."
print(detoxify(test_sentence))

# Test the function with a new sentence
test_sentence = "Trying to keep me fucking drugged so I don't know what's going on."
print(detoxify(test_sentence))

# Test the function with a new sentence
test_sentence = "Damn,man,i wanted the old football coach,man!"
print(detoxify(test_sentence))

from transformers import pipeline
classifier = pipeline('sentiment-analysis', model='distilbert-base-uncased-finetuned-sst-2-english')

def is_toxic(text_to_analyze, comments = False):
  # Run the classifier (model and tokenizer)
  results = classifier(text_to_analyze)

  # Interpret the results
  for result in results:
      label = result['label']
      score = result['score']

      # Heuristic for toxicity based on negative sentiment score
      res = label == 'NEGATIVE' and score > TOXICITY_THRESHOLD
      if comments:
        print(f"Label: {label}, Score: {score}")
      return res

import random
data_test = pd.read_csv(DATA_PATH + '/interim/test.csv')
test_texts = data_test['reference'].tolist()
random.shuffle(test_texts)
test_texts = test_texts[:20]

not_toxic = 0

for toxic_text in test_texts:
    text_to_analyze =detoxify(toxic_text)
    toxic = is_toxic(text_to_analyze, False)
    if not toxic:
      not_toxic += 1
print("\nAccuracy:", not_toxic/len(test_texts))

from datasets import load_metric
metric = load_metric("sacrebleu")

data_test = pd.read_csv(DATA_PATH + 'interim/test.csv').head(100)
test_texts_reference = data_test['reference'].tolist()
test_texts_translation = [[data] for data in data_test['translation'].tolist()]

predictions = [detoxify(toxic_text) for toxic_text in test_texts_reference]
metric.compute(predictions = predictions, references=test_texts_translation)
