==> notebooks/5.0-BERT-model-pre-trained.ipynb <==
!pip install transformers[torch]

import torch
from torch.utils.data import Dataset, DataLoader
from transformers import BertTokenizer, BertForSequenceClassification
from transformers import AdamW
import pandas as pd
from tqdm import tqdm


# Load the dataset
df = pd.read_csv('most_toxic_data.csv').head(1000)

# Tokenization
tokenizer = BertTokenizer.from_pretrained('bert-base-uncased')
tokens = tokenizer(df['reference'].tolist(), max_length=128, padding=True, truncation=True, return_tensors="pt")

# Dataset class
class DetoxificationDataset(Dataset):
    def __init__(self, encodings, labels):
        self.encodings = encodings
        self.labels = labels

    def __getitem__(self, idx):
        item = {key: torch.tensor(val[idx]) for key, val in self.encodings.items()}
        item['labels'] = torch.tensor(int(self.labels[idx]))
        return item

    def __len__(self):
        return len(self.labels)

# Prepare dataset and dataloader
dataset = DetoxificationDataset(tokens, df['trn_tox'])
dataloader = DataLoader(dataset, batch_size=16, shuffle=True)

# Load pre-trained BERT model
model = BertForSequenceClassification.from_pretrained('bert-base-uncased')

# Define optimizer
optimizer = AdamW(model.parameters(), lr=1e-5)
epochs = 5


# Assuming 'epochs' is defined and you have a 'dataloader' for your data
loss_values = []  # List to store loss values

for epoch in tqdm(range(epochs), desc="Epoch"):
    total_loss = 0  # Variable to store the epoch loss

    for batch in tqdm(dataloader, desc="Iteration"):
        optimizer.zero_grad()

        # Unpack the batch contents and feed them into the model
        input_ids = batch['input_ids']
        attention_mask = batch['attention_mask']
        labels = batch['labels']

        # Forward pass and compute loss
        outputs = model(input_ids, attention_mask=attention_mask, labels=labels)
        loss = outputs.loss

        # Backward pass
        loss.backward()
        optimizer.step()

        # Update total loss
        total_loss += loss.item()

    # Calculate the average loss over the entire epoch
    avg_epoch_loss = total_loss / len(dataloader)
    print(f"Average Loss at Epoch {epoch}: {avg_epoch_loss}")

    # Store the average loss value for plotting or analysis later
    loss_values.append(avg_epoch_loss)

# After training loop, you might want to save the model


# Save the model's state_dict
model.save_pretrained('bert_model')
tokenizer.save_pretrained('bert_model')

!zip -r bert_model.zip bert_model

def predict(toxicity_classifier, tokenizer, text):
    # Prepare the input text
    inputs = tokenizer(text, return_tensors="pt", padding=True, truncation=True, max_length=128)
    input_ids = inputs['input_ids']
    attention_mask = inputs['attention_mask']

    # Get model predictions
    with torch.no_grad():
        outputs = toxicity_classifier(input_ids, attention_mask=attention_mask)

    logits = outputs.logits.squeeze()  # Squeeze the logits to remove any extra dimensions

    # Apply softmax if there are multiple classes (logits would be a 1D Tensor with more than one element)
    if logits.numel() > 1:  # More than one element means multi-class classification
        probabilities = torch.softmax(logits, dim=0)
        predicted_class = torch.argmax(probabilities).item()
        return probabilities, predicted_class
    # Apply sigmoid if it's binary classification (logits would be a single-element tensor)
    else:  # Single element means binary classification
        probability = torch.sigmoid(logits).item()
        return probability

example_text = "I like that shit."
prediction = predict(model, tokenizer, example_text)

# Process the prediction
if isinstance(prediction, tuple):  # Multi-class classification
    probabilities, predicted_class = prediction
    print(f"Probabilities: {probabilities[0].item()}")
    print(f"Predicted toxic: {'Yes' if predicted_class == 0 else 'No'}")
